# Fix slow dot in numba